Refactor sending email algorithm from payTransLoader to payTransMailer (Emailer Progress loader)
